added livereload server

+++ Gruntfile.js
@@ -10,6 +10,12 @@ module.exports = function(grunt) {
 			}
 		},
 		watch: {
+			watchHtml: {
+				files: ['*.html', '*.js', '*.css'],
+				options: {
+					livereload: true
+				}
+			},
 			mdAdded: {
 				files: ['*.md'],
 				tasks: ['markdown'],
@@ -27,15 +33,25 @@ module.exports = function(grunt) {
 				}				
 			}
 		},
+		connect: {
+			server: {
+				options: {
+					port: 8000,
+					base: '.',
+					livereload: true
+				}
+			}
+		},
 	});
 
 	grunt.loadNpmTasks('grunt-contrib-watch');
 	grunt.loadNpmTasks('grunt-markdown');
+	grunt.loadNpmTasks('grunt-contrib-connect');
 
-	grunt.registerTask('default', ['watch']);
+	grunt.registerTask('default', ['connect:server', 'markdown', 'watch']);
 
 	grunt.event.on('watch', function(action, filepath, target) {
-		if (action == 'deleted') 
+		if (action == 'deleted' && filepath.substr(filepath.lastIndexOf('.') + 1) == 'md') 
 			grunt.file.delete(filepath.substr(0, filepath.lastIndexOf('.')) + '.html');
 		else
 			grunt.config('markdown.all.src', filepath);
